Fixes for Float32 and Float64 to run simultaneously

+++ src/core_types.jl
@@ -125,7 +125,7 @@ struct SingleCUDAContext <: GPUTervContext
     block_size
     device
     matrix_layout
-    function SingleCUDAContext(float_t::Type = Float32, index_t::Type = Int32, block_size = 256, layout = EquationMajorLayout())
+    function SingleCUDAContext(float_t::Type = Float32, index_t::Type = Int64, block_size = 256, layout = EquationMajorLayout())
         @assert CUDA.functional() "CUDA must be functional for this context."
         return new(float_t, index_t, block_size, CUDADevice(), layout)
     end
